Remove mrmadjb.github.io from piracy list (#41)
This repository never contained piracy, and does not seem to be dangerous.
Package analysis:
com.mrmadjb.*cht: Locale files (/Library/PreferenceBundles/*.bundle/zh-Hant-TW.lproj) for some tweaks. Does not actually contain the tweaks.

com.mrmadjb.PxMartPRO: Appears to be a jailbreak detection bypass. Google Search suggested that this is (one of) the original hosting repo. (Website of developer links to a post of the blog of this repo)

com.mrmadjb.esunbankbypass: Also appears to be a jailbreak detection bypass. Google Search suggested that this is (one of) the original hosting repo. (Every search result links to this repo if linked a repo at all + proper credits + presumption of innocence)